Move selectivity optional filter definitions to new files and fix includes
