+++ src/include/duckdb/planner/filter/table_filter_functions.hpp
@@ -15,6 +15,7 @@
 #include "duckdb/common/types/value.hpp"
 #include "duckdb/function/scalar/tablefilter_functions.hpp"
 #include "duckdb/function/scalar_function.hpp"
+#include "duckdb/planner/table_filter_state.hpp"
 
 namespace duckdb {
 
@@ -23,7 +24,43 @@ class Expression;
 class PerfectHashJoinExecutor;
 class PrefixRangeFilter;
 struct DynamicFilterData;
-enum class SelectivityOptionalFilterType : uint8_t;
+
+struct SelectivityOptionalFilterState final : public TableFilterState {
+	enum class FilterStatus { ACTIVE, PAUSED_DUE_TO_HIGH_SELECTIVITY };
+
+	struct SelectivityStats {
+		SelectivityStats(idx_t n_vectors_to_check, float selectivity_threshold);
+
+		void Update(idx_t accepted, idx_t processed);
+		bool IsActive() const;
+		double GetSelectivity() const;
+
+		//! Configuration
+		const idx_t n_vectors_to_check;
+		const float selectivity_threshold;
+
+		//! For computing selectivity stats
+		idx_t tuples_accepted;
+		idx_t tuples_processed;
+		idx_t vectors_processed;
+
+		//! Whether currently paused
+		FilterStatus status;
+
+		//! For increasing pause if filter is not selective enough
+		idx_t pause_multiplier;
+	};
+
+	unique_ptr<TableFilterState> child_state;
+	SelectivityStats stats;
+
+	explicit SelectivityOptionalFilterState(unique_ptr<TableFilterState> child_state, const idx_t n_vectors_to_check,
+	                                        const float selectivity_threshold)
+	    : child_state(std::move(child_state)), stats(n_vectors_to_check, selectivity_threshold) {
+	}
+};
+
+enum class SelectivityOptionalFilterType : uint8_t { MIN_MAX, BF, PHJ, PRF };
 
 void GetThresholdAndVectorsToCheck(SelectivityOptionalFilterType type, float &selectivity_threshold,
                                    idx_t &n_vectors_to_check);
