+++ src/planner/filter/table_filter_selectivity_optional_function.cpp
@@ -16,6 +16,45 @@
 
 namespace duckdb {
 
+SelectivityOptionalFilterState::SelectivityStats::SelectivityStats(const idx_t n_vectors_to_check,
+                                                                   const float selectivity_threshold)
+    : n_vectors_to_check(n_vectors_to_check), selectivity_threshold(selectivity_threshold), tuples_accepted(0),
+      tuples_processed(0), vectors_processed(0), status(FilterStatus::ACTIVE), pause_multiplier(0) {
+}
+
+void SelectivityOptionalFilterState::SelectivityStats::Update(idx_t accepted, idx_t processed) {
+	vectors_processed++;
+	tuples_accepted += accepted;
+	tuples_processed += processed;
+
+	static constexpr idx_t VECTOR_PAUSE = 10;
+	D_ASSERT(n_vectors_to_check < VECTOR_PAUSE);
+	if (vectors_processed == MaxValue<idx_t>(pause_multiplier, 1) * VECTOR_PAUSE) {
+		vectors_processed = 0;
+		tuples_accepted = 0;
+		tuples_processed = 0;
+		status = FilterStatus::ACTIVE;
+	} else if (vectors_processed >= n_vectors_to_check) {
+		// pause the filter if we processed enough vectors and the selectivity is too high
+		if (GetSelectivity() >= selectivity_threshold) {
+			status = FilterStatus::PAUSED_DUE_TO_HIGH_SELECTIVITY;
+			pause_multiplier++; // increase the pause duration
+		} else {
+			pause_multiplier = 0; // selective enough, reset the pause duration
+		}
+	}
+}
+
+bool SelectivityOptionalFilterState::SelectivityStats::IsActive() const {
+	return status == FilterStatus::ACTIVE;
+}
+double SelectivityOptionalFilterState::SelectivityStats::GetSelectivity() const {
+	if (tuples_processed == 0) {
+		return 0.0;
+	}
+	return static_cast<double>(tuples_accepted) / static_cast<double>(tuples_processed);
+}
+
 SelectivityOptionalFilterFunctionData::SelectivityOptionalFilterFunctionData(unique_ptr<Expression> child_filter_expr_p,
                                                                              float selectivity_threshold_p,
                                                                              idx_t n_vectors_to_check_p)
